Removed protocol relative URL for jQuery
Nowadays the tendency of the web is to move to HTTPS, so the use of protocol-relative URLs has become an anti-pattern therefor I make this commit to update the codebase to reflect this. 

Furthermore, I have added charset="utf-8" and type="text/javascript" to tag_css and tag_js as to follow best practice.